% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boto3.R
\name{assert_boto3_available}
\alias{assert_boto3_available}
\title{Fail if boto3 Python module is not installed}
\usage{
assert_boto3_available()
}
\description{
Fail if boto3 Python module is not installed
}
\keyword{internal}
