% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base64_dec}
\alias{base64_dec}
\title{Base64-decode a string into raw bytes using Python's base64 module}
\usage{
base64_dec(text)
}
\arguments{
\item{text}{string}
}
\value{
\code{raw} bytes
}
\description{
Base64-decode a string into raw bytes using Python's base64 module
}
\examples{
\dontrun{
botor:::base64_dec(botor:::base64_enc(charToRaw('foobar')))
}
}
\seealso{
\code{\link{base64_enc}}
}
\keyword{internal}
