% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base64_enc}
\alias{base64_enc}
\title{Base64-encode raw bytes using Python's base64 module}
\usage{
base64_enc(text)
}
\arguments{
\item{text}{\code{raw}, R string or Python string}
}
\value{
string
}
\description{
Base64-encode raw bytes using Python's base64 module
}
\seealso{
\code{\link{base64_dec}}
}
\keyword{internal}
