% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boto3.R
\docType{data}
\name{boto3}
\alias{boto3}
\title{Raw access to the boto3 module imported at package load time}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
boto3
}
\description{
Raw access to the boto3 module imported at package load time
}
\note{
You may rather want to use \code{\link{botor}} instead, that provides a fork-safe \code{boto3} session.
}
\keyword{datasets}
