% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{botor}
\alias{botor}
\title{Raw access to the boto3 module}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
botor
}
\description{
Raw access to the boto3 module
}
\keyword{datasets}
