% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{botor_client}
\alias{botor_client}
\title{Creates an initial or reinitialize an already existing AWS client}
\usage{
botor_client(service)
}
\arguments{
\item{service}{string, eg S3 or IAM}
}
\value{
cached AWS client
}
\description{
Creates an initial or reinitialize an already existing AWS client
}
\keyword{internal}
