% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botor.R
\docType{data}
\name{botor_session}
\alias{botor_session}
\title{Internal boto3 session}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
botor_session
}
\description{
Internal boto3 session
}
\keyword{internal}
