% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botor.R
\name{botor_session_id}
\alias{botor_session_id}
\title{Look up internal Python id of the Boto3 session}
\usage{
botor_session_id()
}
\value{
int
}
\description{
Look up internal Python id of the Boto3 session
}
\keyword{internal}
