% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botor.R
\name{botor_session_pid}
\alias{botor_session_pid}
\title{Look up the PID used to initialize the Boto3 session}
\usage{
botor_session_pid()
}
\value{
int
}
\description{
Look up the PID used to initialize the Boto3 session
}
\keyword{internal}
