% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botor.R
\name{botor_session_uuid}
\alias{botor_session_uuid}
\title{Look up the UUID of the initialized Boto3 session}
\usage{
botor_session_uuid()
}
\value{
int
}
\description{
Look up the UUID of the initialized Boto3 session
}
\keyword{internal}
