% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_s3_path}
\alias{check_s3_path}
\title{Check if an argument looks like an S3 bucket}
\usage{
check_s3_path(x)
}
\arguments{
\item{x}{string}
}
\description{
Check if an argument looks like an S3 bucket
}
\examples{
\dontrun{
check_s3_path('s3://foo/bar')
check_s3_path('https://foo/bar')
assert_s3_path('https://foo/bar')
}
}
