% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_bytes_literals_to_string}
\alias{coerce_bytes_literals_to_string}
\title{Transforms a python2 string literal or python3 bytes literal into an R string}
\usage{
coerce_bytes_literals_to_string(x)
}
\arguments{
\item{x}{string}
}
\value{
string
}
\description{
This is useful to call eg for the KMS call, where python2 returns a string, but python3 returns bytes literals -- calling "decode" is tricky, but bytearray conversion, then passing the raw vector to R and converting that a string works.
}
\keyword{internal}
