% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam.R
\name{iam}
\alias{iam}
\title{The default, fork-safe IAM client on the top of \code{botor}}
\usage{
iam()
}
\value{
\code{botocore.client.IAM}
}
\description{
The default, fork-safe IAM client on the top of \code{botor}
}
\references{
\url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/iam.html}
}
