% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam.R
\name{iam_get_user}
\alias{iam_get_user}
\title{Retrieves information about the specified IAM user, including the user's creation date, path, unique ID, and ARN}
\usage{
iam_get_user(...)
}
\arguments{
\item{...}{optional extra arguments passed}
}
\value{
list
}
\description{
Retrieves information about the specified IAM user, including the user's creation date, path, unique ID, and ARN
}
\references{
\url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/iam.html#IAM.Client.get_user}
}
