% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis.R
\name{kinesis}
\alias{kinesis}
\title{The default, fork-safe Kinesis client on the top of \code{botor}}
\usage{
kinesis()
}
\value{
\code{botocore.client.Kinesis}
}
\description{
The default, fork-safe Kinesis client on the top of \code{botor}
}
\references{
\url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/kinesis.html}
}
