% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis.R
\name{kinesis_describe_stream}
\alias{kinesis_describe_stream}
\title{Describes the specified Kinesis data stream}
\usage{
kinesis_describe_stream(stream)
}
\arguments{
\item{stream}{the name of the stream to describe}
}
\value{
list
}
\description{
Describes the specified Kinesis data stream
}
\references{
\url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/kinesis.html#Kinesis.Client.describe_stream}
}
