% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{kms}
\alias{kms}
\title{The default, fork-safe KMS client on the top of \code{botor}}
\usage{
kms()
}
\value{
\code{s3.ServiceResource}
}
\description{
The default, fork-safe KMS client on the top of \code{botor}
}
\references{
\url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#service-resource}
}
