% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{kms_decrypt}
\alias{kms_decrypt}
\title{Decrypt cipher into plain text via KMS}
\usage{
kms_decrypt(cipher, simplify = TRUE)
}
\arguments{
\item{cipher}{Base64-encoded ciphertext}

\item{simplify}{returns decrypted plain-text instead of raw list}
}
\value{
decrypted text as string or \code{list}
}
\description{
Decrypt cipher into plain text via KMS
}
\seealso{
\code{\link{kms_encrypt}}
}
