% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mime_guess}
\alias{mime_guess}
\title{Guess the type of a file based on the filename using \code{mimetypes} Python module}
\usage{
mime_guess(file)
}
\arguments{
\item{file}{path}
}
\value{
string
}
\description{
Guess the type of a file based on the filename using \code{mimetypes} Python module
}
