% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{python_version}
\alias{python_version}
\title{Looks up major version of Python (eg 2 or 3)}
\usage{
python_version()
}
\value{
number
}
\description{
Looks up major version of Python (eg 2 or 3)
}
\keyword{internal}
