% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_modules.R
\name{require_python_builtins}
\alias{require_python_builtins}
\title{Imports and caches a Python module}
\usage{
require_python_builtins()
}
\value{
imported Python module
}
\description{
Imports and caches a Python module
}
\keyword{internal}
