% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_modules.R
\name{require_python_module}
\alias{require_python_module}
\title{Imports and caches a Python module}
\usage{
require_python_module(module)
}
\arguments{
\item{module}{a Python module name}
}
\value{
imported Python module
}
\description{
Imports and caches a Python module
}
\keyword{internal}
