% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\docType{data}
\name{s3}
\alias{s3}
\title{Raw access to boto3.resource('s3')}
\format{An object of class \code{NULL} of length 0.}
\usage{
s3
}
\description{
Raw access to boto3.resource('s3')
}
\references{
\url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#service-resource}
}
\keyword{datasets}
