% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_copy}
\alias{s3_copy}
\title{Copy an object from one S3 location to another}
\usage{
s3_copy(uri_source, uri_target)
}
\arguments{
\item{uri_source}{string, location of the source file}

\item{uri_target}{string, location of the target file}
}
\value{
invisibly \code{uri_target}
}
\description{
Copy an object from one S3 location to another
}
\references{
\url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#S3.Object.copy}
}
