% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_delete}
\alias{s3_delete}
\title{Delete an object stored in S3}
\usage{
s3_delete(uri)
}
\arguments{
\item{uri}{string, URI of an S3 object, should start with \code{s3://}, then bucket name and object key}
}
\description{
Delete an object stored in S3
}
