% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_download}
\alias{s3_download}
\title{Download a file from S3}
\usage{
s3_download(object, file, force = TRUE)
}
\arguments{
\item{object}{string, remote S3 path of file to download, should start with 's3://'}

\item{file}{string, location of local file}

\item{overwrite}{boolean, overwrite local file if exists}
}
\value{
invisibly \code{file}
}
\description{
Download a file from S3
}
\examples{
\dontrun{
s3_download('s3://botor/example-data/mtcars.csv', tempfile())
}
}
