% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_exists}
\alias{s3_exists}
\title{Checks if an object exists in S3}
\usage{
s3_exists(uri)
}
\arguments{
\item{uri}{string, URI of an S3 object, should start with \code{s3://}, then bucket name and object key}
}
\value{
boolean
}
\description{
Checks if an object exists in S3
}
\examples{
\dontrun{
s3_exists('s3://botor/example-data/mtcars.csv')
s3_exists('s3://botor/example-data/UNDEFINED.CSVLX')
}
}
