% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_extract_path}
\alias{s3_extract_path}
\title{Extract bucket name and key from S3 path}
\usage{
s3_extract_path(path)
}
\arguments{
\item{path}{S3 path starting with \code{s3://}, bucket name and object key}
}
\value{
list
}
\description{
Extract bucket name and key from S3 path
}
