% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_list_buckets}
\alias{s3_list_buckets}
\title{List all S3 buckets}
\usage{
s3_list_buckets(simplify = TRUE)
}
\arguments{
\item{simplify}{return bucket names as a character vector}
}
\value{
\code{list} of \code{boto3.resources.factory.s3.Bucket} or a character vector
}
\description{
List all S3 buckets
}
