% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_ls}
\alias{s3_ls}
\title{List objects at an S3 path}
\usage{
s3_ls(uri)
}
\arguments{
\item{uri}{string, should start with \code{s3://}, then bucket name and object key prefix}
}
\value{
\code{data.frame} with bucket name, key/path, size, owner, last modification timestamp
}
\description{
List objects at an S3 path
}
