% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_object}
\alias{s3_object}
\title{Create an S3 Object reference from an URI}
\usage{
s3_object(uri)
}
\arguments{
\item{uri}{string, URI of an S3 object, should start with \code{s3://}, then bucket name and object key}
}
\value{
\code{s3$Object}
}
\description{
Create an S3 Object reference from an URI
}
