% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_put_object_tagging}
\alias{s3_put_object_tagging}
\title{Sets tags on s3 object overwriting all existing tags. Note: tags and metadata tags are not the same.
Calls this function under the hood: https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#S3.Client.put_object_tagging}
\usage{
s3_put_object_tagging(uri, tags)
}
\arguments{
\item{uri}{string, URI of an S3 object, should start with \code{s3://}, then bucket name and object key}

\item{tags}{string, of the format list('my_first_key' = 'my_first_value', 'my_second_key' = 'my_second_value') where keys are the tag names and values are the tag values. Keys and values must be strings.}
}
\description{
Sets tags on s3 object overwriting all existing tags. Note: tags and metadata tags are not the same.
Calls this function under the hood: https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#S3.Client.put_object_tagging
}
