% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_read}
\alias{s3_read}
\title{Download and read a file from S3, then clean up}
\usage{
s3_read(object, fun, ...)
}
\arguments{
\item{object}{string, remote S3 path of file to download, should start with 's3://'}

\item{fun}{R function to read the file, eg \code{fromJSON}, \code{fread} or \code{readRDS}}
}
\value{
R object
}
\description{
Download and read a file from S3, then clean up
}
\examples{
\dontrun{
s3_read('s3://botor/example-data/mtcars.csv', read.csv)
}
}
