% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_split_uri}
\alias{s3_split_uri}
\title{Split the bucket name and object key from the S3 URI}
\usage{
s3_split_uri(uri)
}
\arguments{
\item{uri}{string, URI of an S3 object, should start with \code{s3://}, then bucket name and object key}
}
\value{
list
}
\description{
Split the bucket name and object key from the S3 URI
}
\keyword{internal}
