% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_upload_file}
\alias{s3_upload_file}
\title{Upload a file to S3}
\usage{
s3_upload_file(file, uri)
}
\arguments{
\item{file}{string, location of local file}

\item{uri}{string, URI of an S3 object, should start with \code{s3://}, then bucket name and object key}

\item{overwrite}{boolean, overwrite local file if exists}
}
\value{
invisibly \code{uri}
}
\description{
Upload a file to S3
}
\examples{
\dontrun{
write.csv(mtcars, '/tmp/mtcars.csv', row.names = FALSE)
s3_upload_file('/tmp/mtcars.csv', 's3://botor/example-data/mtcars.csv')
}
}
\references{
\url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#S3.Client.upload_file}
}
\seealso{
\code{\link{s3_download_file}}
}
