% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_write}
\alias{s3_write}
\title{Write an R object into S3}
\usage{
s3_write(x, fun, uri, ...)
}
\arguments{
\item{x}{R object}

\item{fun}{R function with \code{file} argument to write \code{x} to disk before uploading, eg \code{write.csv}, \code{write_json} or \code{saveRDS}}

\item{uri}{string, URI of an S3 object, should start with \code{s3://}, then bucket name and object key}

\item{...}{optional further arguments passed to \code{fun}}
}
\description{
Write an R object into S3
}
\note{
The temp file used for this operation is automatically removed.
}
\examples{
\dontrun{
s3_write(mtcars, write.csv, 's3://botor/example-data/mtcars.csv', row.names = FALSE)
s3_write(mtcars, write.csv2, 's3://botor/example-data/mtcars.csv2', row.names = FALSE)
s3_write(mtcars, jsonlite::write_json, 's3://botor/example-data/mtcars.json', row.names = FALSE)
s3_write(mtcars, saveRDS, 's3://botor/example-data/mtcars.RDS')
}
}
