% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm}
\alias{sm}
\title{The default, fork-safe AWS Systems Manager (SecretManager) client on the top of \code{botor}}
\usage{
sm()
}
\value{
\code{botocore.client.secretsmanager}
}
\description{
The default, fork-safe AWS Systems Manager (SecretManager) client on the top of \code{botor}
}
\references{
\url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/secretsmanager.html}
}
