% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_get_secret}
\alias{sm_get_secret}
\title{Read AWS System Manager's Secrets Manager via Secret Manager}
\usage{
sm_get_secret(path, key = NULL, parse_json = TRUE)
}
\arguments{
\item{path}{name/path of the key to be read}

\item{key}{single key or a vector of keys.}

\item{parse_json}{logical. Default TRUE}
}
\value{
(optionally decrypted) value
}
\description{
Read AWS System Manager's Secrets Manager via Secret Manager
}
