% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{sm_get_secrets}
\alias{sm_get_secrets}
\title{Read AWS System Manager's Secrets Manager via Secret Manager}
\usage{
sm_get_secrets(path, key = NULL)
}
\arguments{
\item{path}{name/path of the key to be read}

\item{key}{keyID}
}
\value{
(optionally decrypted) value
}
\description{
Read AWS System Manager's Secrets Manager via Secret Manager
}
