% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm}
\alias{ssm}
\title{The default, fork-safe AWS Systems Manager (SSM) client on the top of \code{botor}}
\usage{
ssm()
}
\value{
\code{botocore.client.SSM}
}
\description{
The default, fork-safe AWS Systems Manager (SSM) client on the top of \code{botor}
}
\references{
\url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ssm.html}
}
