% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm_get_secrets}
\alias{ssm_get_secrets}
\title{Read AWS System Manager's Secrets Manager via Parameter Store}
\usage{
ssm_get_secrets(path, key = NULL, decrypt = TRUE)
}
\arguments{
\item{path}{name/path of the key to be read}

\item{key}{keyID}

\item{decrypt}{decrypt the value or return the raw ciphertext}
}
\value{
(optionally decrypted) value
}
\description{
Read AWS System Manager's Secrets Manager via Parameter Store
}
