% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam.R
\name{sts_whoami}
\alias{sts_whoami}
\title{Returns details about the IAM user or role whose credentials are used to call the operation}
\usage{
sts_whoami()
}
\value{
\code{list} with \code{UserId}, \code{Account} and \code{Arn}
}
\description{
Returns details about the IAM user or role whose credentials are used to call the operation
}
\references{
\url{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/sts.html#STS.Client.get_caller_identity}
}
\seealso{
\code{\link{iam_whoami}}
}
