% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trypy}
\alias{trypy}
\title{Extract error message from a Python exception}
\usage{
trypy(expression)
}
\arguments{
\item{expression}{R expression}
}
\value{
error
}
\description{
Extract error message from a Python exception
}
\examples{
\dontrun{
trypy(botor()$resource('foobar'))
trypy(sum(1:2))
trypy(sum(1:1foo))
}
}
\keyword{internal}
