% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botor.R
\name{botor}
\alias{botor}
\title{The default, fork-safe Boto3 session}
\usage{
botor(aws_access_key_id, aws_secret_access_key, aws_session_token,
  region_name, botocore_session, profile_name)
}
\arguments{
\item{aws_access_key_id}{AWS access key ID}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS temporary session token}

\item{region_name}{Default region when creating new connections}

\item{botocore_session}{Use this Botocore session instead of creating a new default one}

\item{profile_name}{The name of a profile to use. If not given, then the default profile is used}
}
\value{
boto3 \code{Session}
}
\description{
The default, fork-safe Boto3 session
}
