% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botor.R
\name{botor_client}
\alias{botor_client}
\title{Creates an initial or reinitialize an already existing AWS client or resource cached in the package's namespace}
\usage{
botor_client(service, type = c("client", "resource"))
}
\arguments{
\item{service}{string, eg S3 or IAM}

\item{type}{client or resource to be created}
}
\value{
cached AWS client
}
\description{
Creates an initial or reinitialize an already existing AWS client or resource cached in the package's namespace
}
