% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_s3_uri}
\alias{check_s3_uri}
\alias{assert_s3_uri}
\alias{test_s3_uri}
\alias{expect_s3_uri}
\title{Check if an argument looks like an S3 bucket}
\usage{
check_s3_uri(x)
}
\arguments{
\item{x}{string, URI of an S3 object, should start with \code{s3://}, then bucket name and object key}
}
\description{
Check if an argument looks like an S3 bucket
}
\examples{
check_s3_uri('s3://foo/bar')
check_s3_uri('https://foo/bar')
\dontrun{
assert_s3_uri('https://foo/bar')
}
}
