% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm_get_parameter}
\alias{ssm_get_parameter}
\title{Read AWS System Manager's Parameter Store}
\usage{
ssm_get_parameter(path, decrypt = TRUE)
}
\arguments{
\item{path}{name/path of the key to be read}

\item{decrypt}{decrypt the value or return the raw ciphertext}
}
\value{
(optionally decrypted) value
}
\description{
Read AWS System Manager's Parameter Store
}
